adjusted a few table ID references to bigInteger

+++ database/migrations/2021_03_29_143721_create_actionable_items_table.php
@@ -16,7 +16,7 @@ class CreateActionableItemsTable extends Migration
         Schema::create('actionable_items', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->integer('created_by');
+            $table->bigInteger('created_by');
             $table->string('topic_status');
             $table->integer('list_order');
             $table->timestamps();
