+++ database/migrations/2021_03_29_144206_create_users_viewed_topic_table.php
@@ -15,8 +15,8 @@ class CreateUsersViewedTopicTable extends Migration
     {
         Schema::create('users_viewed_topic', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('topic_id');
+            $table->bigInteger('user_id');
+            $table->bigInteger('topic_id');
             $table->timestamps();
         });
     }
