+++ database/migrations/2021_03_29_144254_create_users_viewed_comment_table.php
@@ -15,8 +15,8 @@ class CreateUsersViewedCommentTable extends Migration
     {
         Schema::create('users_viewed_comment', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('comment_id');
+            $table->bigInteger('user_id');
+            $table->bigInteger('comment_id');
             $table->timestamps();
         });
     }
